backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
tavily-python>=0.3.0
langgraph>=0.0.20
boto3>=1.34.0
python-dotenv>=1.0.0
arxiv>=2.1.0
supabase>=2.0.0